Moved query parameter handling to client side event handler; record mapping only uses attachments from now on. Corruption detection is also considered query param handling.
