+++ src/main/java/io/divolte/server/IncomingRequestProcessor.java
@@ -35,13 +35,8 @@ import io.divolte.server.recordmapping.UserAgentParserAndCache;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.util.AttachmentKey;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Deque;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.SortedMap;
-import java.util.TreeMap;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -51,8 +46,6 @@ import org.apache.avro.generic.GenericRecord;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.common.hash.HashFunction;
-import com.google.common.hash.Hashing;
 import com.typesafe.config.Config;
 
 @ParametersAreNonnullByDefault
@@ -150,8 +143,7 @@ public final class IncomingRequestProcessor implements ItemProcessor<HttpServerE
 
     @Override
     public ProcessingDirective process(final HttpServerExchange exchange) {
-        final boolean corrupt = !isRequestChecksumCorrect(exchange);
-        exchange.putAttachment(CORRUPT_EVENT_KEY, corrupt);
+        final boolean corrupt = exchange.getAttachment(CORRUPT_EVENT_KEY);
 
         if (!corrupt || keepCorrupted) {
             final CookieValue party = exchange.getAttachment(PARTY_COOKIE_KEY);
@@ -195,62 +187,5 @@ public final class IncomingRequestProcessor implements ItemProcessor<HttpServerE
         if (null != hdfsFlushingPool) {
             hdfsFlushingPool.enqueue(avroBuffer.getPartyId().value, avroBuffer);
         }
-    }
-
-    private static final HashFunction CHECKSUM_HASH = Hashing.murmur3_32();
-
-    private static boolean isRequestChecksumCorrect(final HttpServerExchange exchange) {
-        // This is not intended to be robust against intentional tampering; it is intended to guard
-        // against proxies and the like that may have truncated the request.
-
-        return queryParamFromExchange(exchange, CHECKSUM_QUERY_PARAM)
-                .map(ClientSideCookieEventHandler::tryParseBase36Long)
-                .map((expectedChecksum) -> {
-                    /*
-                     * We could optimize this by calculating the checksum directly, instead of building up
-                     * the intermediate string representation. For now the debug value of the string exceeds
-                     * the benefits of going slightly faster.
-                     */
-                    final String canonicalRequestString = buildNormalizedChecksumString(exchange.getQueryParameters());
-                    final int requestChecksum =
-                            CHECKSUM_HASH.hashString(canonicalRequestString, StandardCharsets.UTF_8).asInt();
-                    final boolean isRequestChecksumCorrect = expectedChecksum == requestChecksum;
-                    if (!isRequestChecksumCorrect && logger.isDebugEnabled()) {
-                        logger.debug("Checksum mismatch detected; expected {} but was {} for request string: {}",
-                                Long.toString(expectedChecksum, 36),
-                                Integer.toString(requestChecksum, 36),
-                                canonicalRequestString);
-                    }
-                    return isRequestChecksumCorrect;
-                })
-                .orElse(false);
-    }
-
-    private static String buildNormalizedChecksumString(final Map<String,Deque<String>> queryParameters) {
-        return buildNormalizedChecksumString(queryParameters instanceof SortedMap
-                ? (SortedMap)queryParameters
-                : new TreeMap<>(queryParameters));
-    }
-
-    private static String buildNormalizedChecksumString(final SortedMap<String,Deque<String>> queryParameters) {
-        /*
-         * Build up a canonical representation of the query parameters. The canonical order is:
-         *  1) Sort the query parameters by key, preserving multiple values (and their order).
-         *  2) The magic parameter containing the checksum is discarded.
-         *  3) Build up a string. For each parameter:
-         *     a) Append the parameter name, followed by a '='.
-         *     b) Append each value of the parameter, followed by a ','.
-         *     c) Append a ';'.
-         *  This is designed to be unambiguous in the face of many edge cases.
-         */
-        final StringBuilder builder = new StringBuilder();
-        queryParameters.forEach((name, values) -> {
-            if (!CHECKSUM_QUERY_PARAM.equals(name)) {
-                builder.append(name).append('=');
-                values.forEach((value) -> builder.append(value).append(','));
-                builder.append(';');
-            }
-        });
-        return builder.toString();
     }
 }
